requirements.txt
aioredis==1.1.0
ansible==2.6.4
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.4
cffi==1.11.5
channels==2.1.3
channels-redis==2.3.0
constantly==15.1.0
cryptography==2.3.1
daphne==2.2.2
decorator==4.3.0
Django==2.1.1
django-filter==2.0.0
django-redis==4.9.0
djangorestframework==3.8.2
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Jinja2==2.10
MarkupSafe==1.0
msgpack==0.5.6
paramiko==2.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
PyNaCl==1.2.1
pytz==2018.5
PyYAML==3.13
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
wcwidth==0.1.7
zope.interface==4.5.0
